Add mention of positional arguments.
The existing documentation used the term "free arguments," but
programmers familiar with other argument processing frameworks (e.g.,
Python's `argparse`) may be looking for the term "positional
arguments," which is commonly used.  This adds that, so a quick
textual search of the README will show that `command-arguments` is the
way to get positional arguments.